refactor: separate game info to own file
