+++ src/main_menu.cpp
@@ -50,6 +50,7 @@
 #include "ui.h"
 #include "wcwidth.h"
 #include "worldfactory.h"
+#include "game_info.h"
 
 enum class main_menu_opts : int {
     MOTD = 0,
